* Added Network State Detection on API Service Calls * Added showing of Offline Banner whenever the device is offline * Updated amount data type to Double for more precision in doing computations * Removed all database insertions and updating in between events * Relocated database updates on onStop() to ensure that there will be no concurrent modification issue * Updated the resetting of exchange rate info upon successful balance updates * Fixed the limit detection within the TextField from the SellRowView * Implemented limiting of UI updates on some MainScreenContent Views to improve performance

+++ CurrencyExchange/app/src/main/java/com/application/currency/exchange/data/repository/database/RateRepository.kt
@@ -12,7 +12,7 @@ interface RateRepository {
     fun insertRate(vararg rate: Rate)
     fun deleteRate(vararg rate: Rate)
     fun updateRate(vararg rate: Rate)
-    fun updateRateAmount(currency: String, amount: Float)
+    fun updateRateAmount(currency: String, amount: Double)
     suspend fun getAllRates(): List<Rate>
 }
 
@@ -35,7 +35,7 @@ class RateRepositoryImpl @Inject constructor(private val rateDao: RateDao) : Rat
         }
     }
 
-    override fun updateRateAmount(currency: String, amount: Float) {
+    override fun updateRateAmount(currency: String, amount: Double) {
         CoroutineScope(Dispatchers.IO).launch {
             rateDao.updateRateAmount(currency, amount)
         }
